refactor: move helper functions into helper folder
